correções

+++ backend/components/postReminder.js
@@ -28,10 +28,15 @@ router.post('/new', async (req, res) => {
 router.delete('/delete/:id', async (req, res) => {
     const { id } = req.params;
     try {
+        if (!id) {
+            res.status(404).json({message: "Nenhum ID Informado"});
+            return;
+        }
         await prisma.reminder.delete({ where: { id: id } });
-        res.status(201).json({ message: "Lembrete excluído com sucesso" });
+        res.status(200).json({ message: "Lembrete excluído com sucesso" });
+
     } catch (err) {
-        res.status(500).json({ message: "Ops, algo deu errado ao excluir a lembrança." })
+        res.status(500).json({ message: "Ops, algo deu errado ao excluir o lembrete." });
     }
 })
